Latest updates.  moved code to specific Activities page.
