+++ src/BlazorApp.Bootstrap.Data/Dtos/SelectLitsDto.cs
@@ -8,7 +8,7 @@ namespace BlazorApp.Bootstrap.Data.Dtos
 {
     public class SelectLitsDto
     {
-        public long Value { get; set; }
+        public string Value { get; set; }= string.Empty;
         public string DisplayText { get; set; } = string.Empty;
 
     }
