+++ src/BlazorApp.Bootstrap.Navs/Components/Layout/NavBarSide.razor
@@ -18,6 +18,16 @@
                 <li class="nav-item">
                     <a class="nav-link" href="#">Help</a>
                 </li>
+                <li class="nav-item">
+                    <NavLink class="nav-link" href="activities" Match="NavLinkMatch.All">
+                        <span class="" aria-hidden="true"></span> Activities
+                    </NavLink>
+                </li>
+                <li class="nav-item">
+                    <NavLink class="nav-link" href="activity-types" Match="NavLinkMatch.All">
+                        <span class="" aria-hidden="true"></span> Activity Types
+                    </NavLink>
+                </li>
             </ul>
         </div>
     </div>
