+++ src/BlazorApp.Bootstrap.Navs/Components/Layout/NavBarTop.razor
@@ -33,6 +33,24 @@
                         <li><a class="dropdown-item" href="#">Something else here</a></li>
                     </ul>
                 </li>
+                <li class="nav-item dropdown">
+                    <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                        Actions
+                    </a>
+                    <ul class="dropdown-menu">
+                        <li>
+                            <NavLink class="nav-link" href="activities" Match="NavLinkMatch.All">
+                                <span class="" aria-hidden="true"></span> Activities
+                            </NavLink>
+                        </li>
+                        <li><hr class="dropdown-divider"></li>
+                        <li>
+                            <NavLink class="nav-link" href="activity-types" Match="NavLinkMatch.All">
+                                <span class="" aria-hidden="true"></span> Activity Types
+                            </NavLink>
+                        </li>
+                    </ul>
+                </li>
             </ul>
         </div>
     </div>
