+++ src/BlazorApp.Bootstrap.Navs/Components/Pages/ActivityTypes.razor
@@ -1,0 +1,7 @@
+﻿@page "/activity-types"
+
+<h3>ActivityTypes</h3>
+
+@code {
+
+}
